Added installation instructions for Mac OS X

+++ _downloads/get_rand_array.cpp
@@ -6,7 +6,7 @@ void get_rand_array(double* output_array, int length) {
   
   /* Populate input NumPy array with random numbers */
   for (int i=0; i < length; i++)
-    output_array[i] = ((float) rand()) / RAND_MAX;
+    output_array[i] = ((double) rand()) / RAND_MAX;
   
   return;
 }
